Adding an ini config file.
It is now possible to use the argument --config and specify an ini
config file.
E.g ipxact2rst --srcFile register.xml --destDir=. --config default.ini

Without the config file the behaviour is the same.

For now, in the config file it is possible request not to generate the
unused bits if there is a gap between two fields in a register.

And it is possible to request that enums of fields of 1 bit long are
generated.  Default, there will be no enums generated if the field is only
one bit width.

The testing will test it without config file, with a default config file
and with a config file where the values are none default.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>