A register is allowed not to have a description in the IP-XACT std.
However, the python code was failing when there is no description.
To avoid having to catch at a lot of places in the Python code where
there is no variable reg.desc, I assign an empty string during the parsing
step if not description XML element is there.

I've modified the test input to have one without description.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>